+++ playground/app/app.config.ts
@@ -1,0 +1,8 @@
+export default {
+  ui: {
+    colors: {
+      primary: 'green',
+      neutral: 'zinc',
+    },
+  },
+}