+++ test/fixtures/content/components/content/CustomImage.vue
@@ -1,0 +1,10 @@
+<template>
+  <img
+    :src="src"
+    class="custom-image"
+  >
+</template>
+
+<script setup lang="ts">
+defineProps<{ src: string }>()
+</script>